Redesigns how app tokens are stored
We were not storing tokens on a per-server basis, which would cause
problems for developers talking to more than one cryptid server (say,
dev vs prod) in two different shells. We now store tokens per the server
you're configured to talk to, making it easier to not have to relog all
the time.

This commit also fixes the way we export constants from the cli package.
We now export constants in UPPER_CASE instead opf lowercase.